Check whether the file "common_dependencys.txt" exists before reading this file.

+++ inlong-distribution/script/prepare_common_dependency.sh
@@ -23,7 +23,9 @@ cd "$(dirname "$0")"/../ || exit
 jar_file_num=`ls -l $1 |grep jar |wc -l`
 
 if [ $jar_file_num -eq 0 ]; then
-    for line in $(cat "$1/common_dependencys.txt"); do
-        cp "./lib/$line" $1/
-    done
+    if [ -e $1/common_dependencys.txt ]; then
+        for line in $(cat "$1/common_dependencys.txt"); do
+            cp "./lib/$line" $1/
+        done
+    fi
 fi
